refactor: rename extensions package to util
