+++ app/src/main/java/org/kabiri/android/usbterminal/model/ArduinoDevice.kt
@@ -1,8 +1,8 @@
 package org.kabiri.android.usbterminal.model
 
 import android.hardware.usb.UsbDevice
-import org.kabiri.android.usbterminal.extensions.isCloneArduinoBoard
-import org.kabiri.android.usbterminal.extensions.isOfficialArduinoBoard
+import org.kabiri.android.usbterminal.util.isCloneArduinoBoard
+import org.kabiri.android.usbterminal.util.isOfficialArduinoBoard
 
 internal data class ArduinoDevice(
     val device: UsbDevice,
